refactor: split store api slices into multiple files
