+++ front/src/store/hooks.ts
@@ -1,0 +1,7 @@
+import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
+
+import { AppDispatch, RootState } from './store';
+
+// Store hooks for this app
+export const useAppDispatch: () => AppDispatch = useDispatch;
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
